syscall: reduce complexity of ptrace shim on linux
The syscall package was generating two shims for the __go_ptrace
function on Linux, one named "ptrace" and one named "raw_ptrace", which
differed slightly in the types of their arguments. No need to have two,
so redirect all callers to use one shim named "ptrace", which uses
uintptrs for the addr and data arguments. This hews most closely to the
upstream ptrace wrapper.

Additionally, only compile the __go_ptrace varargs shim on Linux. The C
ptrace function is very much not portable (e.g., NetBSD uses `int data`
instead of `void* data`), and so far Linux is the only platform that
needs the varargs shim.

Change-Id: Ib589ffd504e12440b726490c0a212d9ba570eb2c